refactor: move createCachedRecords to service
